Add README.md and some fixes
Add README.md.
MSRCP: Algorithm modified to avoid color clipping.
MSRCP: Better handling for fulls=False input with out of range values.
Fix potential glitch where size_t is compared with 0.
Fix gcc compiling warnings and errors.
#include modified.

+++ include/MSR.h
@@ -2,7 +2,6 @@
 #define MSR_H_
 
 
-#include <vapoursynth\VapourSynth.h>
 #include "Helper.h"
 
 
@@ -24,7 +23,7 @@ public:
 
     int process[3];
 
-    size_t HistBins = 4096;
+    int HistBins = 4096;
 
 public:
     MSRData(const VSAPI *_vsapi = nullptr)
